+++ src/ArduinoJson/Deserialization/deserialize.hpp
@@ -22,40 +22,37 @@ struct first_or_void<T, Rest...> {
   using type = T;
 };
 
-template <template <typename> class TDeserializer, typename TReader>
-TDeserializer<TReader> makeDeserializer(ResourceManager* resources,
-                                        TReader reader) {
-  ARDUINOJSON_ASSERT(resources != 0);
-  return TDeserializer<TReader>(resources, reader);
+template <template <typename> class TDeserializer, typename TDestination,
+          typename TReader, typename TOptions>
+DeserializationError doDeserialize(TDestination&& dst, TReader reader,
+                                   TOptions options) {
+  auto data = VariantAttorney::getOrCreateData(dst);
+  if (!data)
+    return DeserializationError::NoMemory;
+  auto resources = VariantAttorney::getResourceManager(dst);
+  dst.clear();
+  return TDeserializer<TReader>(resources, reader)
+      .parse(*data, options.filter, options.nestingLimit);
 }
 
-template <template <typename> class TDeserializer, typename TStream,
-          typename... Args,
+template <template <typename> class TDeserializer, typename TDestination,
+          typename TStream, typename... Args,
           typename = typename enable_if<  // issue #1897
               !is_integral<typename first_or_void<Args...>::type>::value>::type>
-DeserializationError deserialize(JsonDocument& doc, TStream&& input,
+DeserializationError deserialize(TDestination&& dst, TStream&& input,
                                  Args... args) {
-  auto reader = makeReader(detail::forward<TStream>(input));
-  auto data = VariantAttorney::getData(doc);
-  auto resources = VariantAttorney::getResourceManager(doc);
-  auto options = makeDeserializationOptions(args...);
-  doc.clear();
-  return makeDeserializer<TDeserializer>(resources, reader)
-      .parse(*data, options.filter, options.nestingLimit);
+  return doDeserialize<TDeserializer>(
+      dst, makeReader(detail::forward<TStream>(input)),
+      makeDeserializationOptions(args...));
 }
 
-template <template <typename> class TDeserializer, typename TChar,
-          typename Size, typename... Args,
+template <template <typename> class TDeserializer, typename TDestination,
+          typename TChar, typename Size, typename... Args,
           typename = typename enable_if<is_integral<Size>::value>::type>
-DeserializationError deserialize(JsonDocument& doc, TChar* input,
+DeserializationError deserialize(TDestination&& dst, TChar* input,
                                  Size inputSize, Args... args) {
-  auto reader = makeReader(input, size_t(inputSize));
-  auto data = VariantAttorney::getData(doc);
-  auto resources = VariantAttorney::getResourceManager(doc);
-  auto options = makeDeserializationOptions(args...);
-  doc.clear();
-  return makeDeserializer<TDeserializer>(resources, reader)
-      .parse(*data, options.filter, options.nestingLimit);
+  return doDeserialize<TDeserializer>(dst, makeReader(input, size_t(inputSize)),
+                                      makeDeserializationOptions(args...));
 }
 
 ARDUINOJSON_END_PRIVATE_NAMESPACE